Inline bindings within raw-instance
